fix: refactor simulation
