+++ index.ts
@@ -4,31 +4,25 @@
 
 require('dotenv').config()
 import fs from 'fs'
-import { DAOs, DAO_NAME, PROPOSAL_FILTER, OMIT_CACHE } from './utils/constants'
+import { DAO_NAME, PROPOSAL_FILTER, OMIT_CACHE, AAVE_GOV_V2_ADDRESS } from './utils/constants'
 import { provider } from './utils/clients/ethers'
-import { simulate } from './utils/clients/tenderly'
-import { AllCheckResults, ProposalData, SimulationConfig, SimulationConfigBase, SimulationData } from './types'
+import { AllCheckResults, ProposalData, SimulationResult } from './types'
 import ALL_CHECKS from './checks'
 import { toProposalReport } from './presentation/markdown'
-import { aaveGovV2, isProposalStateImmutable, PROPOSAL_STATES } from './utils/contracts/aave-governance-v2'
+import { aaveGovernanceContract, isProposalStateImmutable, PROPOSAL_STATES } from './utils/contracts/aave-governance-v2'
 import { executor } from './utils/contracts/executor'
 import { PromisePool } from '@supercharge/promise-pool'
+import { simulateProposal } from './utils/simulations/proposal'
 
 // load cache
 const proposalStateCachePath = './proposal-states.json'
 let cache: { [key: string]: string } = {}
 if (fs.existsSync(proposalStateCachePath)) cache = require(proposalStateCachePath)
 
-// make sure env variables are set
-if (!DAOs[DAO_NAME as keyof typeof DAOs]) {
-  throw new Error(`DAO_NAME:${DAO_NAME} unknown. Must be one of ${Object.keys(DAOs).join(',')}`)
-}
-const AAVE_GOV_V2_ADDRESS = DAOs[DAO_NAME as keyof typeof DAOs]
-
-function getProposalFileName(config: SimulationConfig) {
-  const basePath = `${config.daoName}/${config.governanceAddress}`
+function getProposalFileName(proposalId: number) {
+  const basePath = `${DAO_NAME}/${AAVE_GOV_V2_ADDRESS}`
   // zero padding so the alphabetic file sorting on github works
-  const filename = `${config.proposalId.toString().padStart(3, '0')}.md`
+  const filename = `${proposalId.toString().padStart(3, '0')}.md`
   const dir = `./reports/${basePath}/`
   if (!fs.existsSync(dir)) fs.mkdirSync(dir, { recursive: true })
   return `${dir}/${filename}`
@@ -37,65 +31,46 @@ function getProposalFileName(config: SimulationConfig) {
 async function runSimulation() {
   // --- Run simulations ---
   // Fetch all proposal IDs
-  const governance = aaveGovV2(AAVE_GOV_V2_ADDRESS)
-  const proposalsCount = await governance.getProposalsCount()
+  const proposalsCount = await aaveGovernanceContract.getProposalsCount()
   const _allProposalIds = [...Array(Number(proposalsCount)).keys()]
 
   const allProposalIds = PROPOSAL_FILTER
     ? _allProposalIds.filter((id) => PROPOSAL_FILTER!.includes(id))
     : _allProposalIds
 
-  const states = await Promise.all(allProposalIds.map((id) => governance.getProposalState(id)))
-  const simProposals: { id: number; simType: SimulationConfigBase['type']; state: keyof typeof PROPOSAL_STATES }[] =
-    allProposalIds.map((id, i) => {
-      // If state is `Executed` (state 7), we use the executed sim type and effectively just
-      // simulate the real transaction. For all other states, we use the `proposed` type because
-      // state overrides are required to simulate the transaction
-      const state = String(states[i]) as keyof typeof PROPOSAL_STATES
-      const isExecuted = PROPOSAL_STATES[state] === 'Executed'
-      return { id, simType: isExecuted ? 'executed' : 'proposed', state }
-    })
-  const simProposalsIds = simProposals.map((sim) => sim.id)
-
   // Simulate them
   // We intentionally do not run these in parallel to avoid hitting Tenderly API rate limits or flooding
   // them with requests if we e.g. simulate all proposals for a governance (instead of just active ones)
-  const numProposals = simProposals.length
-  console.log(`Simulating ${numProposals} ${DAO_NAME} proposals: IDs of ${JSON.stringify(simProposalsIds)}`)
+  console.log(`Simulating ${allProposalIds.length} ${DAO_NAME} proposals: IDs of ${JSON.stringify(allProposalIds)}`)
 
   const errors: any[] = []
-  const { results: _results } = await PromisePool.for(simProposals)
+  const { results: _results } = await PromisePool.for(allProposalIds)
     .withConcurrency(5)
-    .handleError(async (error, simProposal) => {
+    .handleError(async (error, proposalId) => {
       errors.push(error)
-      console.log(`error simulating proposal ${simProposal.id}`)
+      console.log(`error simulating proposal ${proposalId}`)
       throw error
     })
-    .process(async (simProposal) => {
-      console.log(`  Simulating ${DAO_NAME} proposal ${simProposal.id}...`)
-      const config: SimulationConfig = {
-        type: simProposal.simType,
-        daoName: DAO_NAME,
-        governanceAddress: governance.address,
-        proposalId: simProposal.id,
-        currentProposalState: simProposal.state,
-      }
+    .process(async (proposalId) => {
+      console.log(`  Simulating ${DAO_NAME} proposal ${proposalId}...`)
       // skip proposals when they were simulated before in an immutable state
-      const skip = isProposalStateImmutable(simProposal.state) && cache[simProposal.id] === simProposal.state
-      if (fs.existsSync(getProposalFileName(config)) && skip && !OMIT_CACHE) {
-        console.log(`Skipped simulation for ${simProposal.id}`)
+      const proposalState = (await aaveGovernanceContract.getProposalState(proposalId)) as keyof typeof PROPOSAL_STATES
+      const skip = isProposalStateImmutable(proposalState) && cache[proposalId] === proposalState
+      if (fs.existsSync(getProposalFileName(proposalId)) && skip && !OMIT_CACHE) {
+        console.log(`Skipped simulation for ${proposalId}`)
       } else {
-        const { sim, proposal, latestBlock } = await simulate(config)
-        return { sim, proposal, latestBlock, config }
+        const { sim, latestBlock, proposal } = await simulateProposal(proposalId)
+        // if (targetsArc(sim)) {
+        //   return [{ sim, proposal, latestBlock, config }, await simulateArc(sim)]
+        // }
+        return { sim, proposal, latestBlock }
       }
     })
   if (errors.length) throw errors
-  return _results.filter((r) => r) as SimulationData[]
+  return _results.flat().filter((r) => r) as SimulationResult[]
 }
 
-async function generateReports(simOutputs: SimulationData[]) {
-  const governance = aaveGovV2(AAVE_GOV_V2_ADDRESS)
-
+async function generateReports(simOutputs: SimulationResult[]) {
   console.log('Starting proposal checks and report generation...')
   const errors: any[] = []
   await PromisePool.for(simOutputs)
@@ -107,8 +82,12 @@ async function generateReports(simOutputs: SimulationData[]) {
     })
     .process(async (simOutput) => {
       // Run checks
-      const { sim, proposal, latestBlock, config } = simOutput
-      const proposalData: ProposalData = { governance, provider, executor: executor(await proposal.executor) }
+      const { sim, proposal, latestBlock } = simOutput
+      const proposalData: ProposalData = {
+        governance: aaveGovernanceContract,
+        provider,
+        executor: executor(proposal.executor),
+      }
       console.log(`  Running for proposal ${proposal.id}...`)
       const checkResults: AllCheckResults = Object.fromEntries(
         await Promise.all(
@@ -135,8 +114,8 @@ async function generateReports(simOutputs: SimulationData[]) {
       )
 
       // save report
-      fs.writeFileSync(getProposalFileName(config), report)
-      cache[proposal.id.toString()] = config.currentProposalState
+      fs.writeFileSync(getProposalFileName(proposal.id.toNumber()), report)
+      cache[proposal.id.toString()] = proposal.state
     })
   if (errors.length) throw errors
 }
