+++ package.json
@@ -29,13 +29,16 @@
   "dependencies": {
     "@aave/contract-helpers": "^1.2.1",
     "@supercharge/promise-pool": "^2.1.0",
+    "bignumber.js": "^9.0.2",
     "bn.js": "^5.2.0",
     "dotenv": "^10.0.0",
     "ethers": "^5.6.7",
     "gray-matter": "^4.0.3",
     "jest": "^28.1.0",
     "micro-ftch": "^0.3.1",
+    "reflect-metadata": "^0.1.13",
     "ts-node": "^10.2.1",
+    "tslib": "^2.4.0",
     "typescript": "^4.4.2"
   },
   "devDependencies": {
